+++ Modules/Microsoft365DSC/DSCResources/MSFT_AADGroup/MSFT_AADGroup.psm1
@@ -45,11 +45,11 @@ function Get-TargetResource
         [System.String]
         $MembershipRuleProcessingState,
 
-        [Parameter()]
+        [Parameter(Mandatory = $true)]
         [System.Boolean]
         $SecurityEnabled,
 
-        [Parameter()]
+        [Parameter(Mandatory = $true)]
         [System.Boolean]
         $MailEnabled,
 
@@ -308,11 +308,11 @@ function Set-TargetResource
         [System.String]
         $MembershipRuleProcessingState,
 
-        [Parameter()]
+        [Parameter(Mandatory = $true)]
         [System.Boolean]
         $SecurityEnabled,
 
-        [Parameter()]
+        [Parameter(Mandatory = $true)]
         [System.Boolean]
         $MailEnabled,
 
@@ -824,11 +824,11 @@ function Test-TargetResource
         [System.String]
         $MembershipRuleProcessingState,
 
-        [Parameter()]
+        [Parameter(Mandatory = $true)]
         [System.Boolean]
         $SecurityEnabled,
 
-        [Parameter()]
+        [Parameter(Mandatory = $true)]
         [System.Boolean]
         $MailEnabled,